refactoring for the PR
