+++ site/content/en/docs/contrib/tests.en.md
@@ -202,9 +202,6 @@ NOTE: DNS forwarding is experimental: https://minikube.sigs.k8s.io/docs/handbook
 #### validateTunnelDelete
 stops `minikube tunnel`
 
-## TestGeneratedLogs
-Verifies logs generated in the /tmp directory
-
 ## TestGuestEnvironment
 verifies files and packges installed inside minikube ISO/Base image
 
@@ -360,4 +357,4 @@ upgrades Kubernetes from oldest to newest
 ## TestMissingContainerUpgrade
 tests a Docker upgrade where the underlying container is missing
 
-TEST COUNT: 115
+TEST COUNT: 114
